hh/src/requirements.txt
pytorch
torch
torchvision
numpy
Pillow